// src/modules/dispute-resolution/enums/timeline-event-type.enum.ts
export enum TimelineEventType {
  DISPUTE_FILED = 'dispute_filed',
  STATUS_CHANGED = 'status_changed',
  EVIDENCE_SUBMITTED = 'evidence_submitted',
  EVIDENCE_VALIDATED = 'evidence_validated',
  ARBITRATOR_ASSIGNED = 'arbitrator_assigned',
  MESSAGE_SENT = 'message_sent',
  DECISION_MADE = 'decision_made',
  APPEAL_FILED = 'appeal_filed',
  ENFORCEMENT_EXECUTED = 'enforcement_executed',
  SLA_BREACH = 'sla_breach',
}